arm/altp2m: Make flush_tlb_domain ready for altp2m.
This commit makes sure that the TLB of a domain considers flushing all
of the associated altp2m views. Therefore, in case a different domain
(not the currently active domain) shall flush its TLBs, the current
implementation loops over all VTTBRs of the different altp2m mappings
per vCPU and flushes the TLBs. This way, a change of one of the altp2m
mapping is considered.  At this point, it must be considered that the
domain --whose TLBs are to be flushed-- is not locked.

Signed-off-by: Sergej Proskurin <proskurin@sec.in.tum.de>
---
Cc: Stefano Stabellini <sstabellini@kernel.org>
Cc: Julien Grall <julien.grall@arm.com>
---
v2: Changed prototype of p2m_load_VTTBR: The struct p2m_domain of the
    idle domain does not hold a back pointer to the domain. Thus,
    p2m_load_VTTBR requires an additional argument of type struct
    domain.  Alternatively, we could adjust the funtion
    arch_domain_create (in ./xen/arch/arm/domain.c) to correctly
    configure the p2m->domain pointer.

    Added function flush_tlb_p2m that takes over responsibility to
    determine whether altp2m is active and flush the TLBs appropriately.
    We did not remove the function flush_tlb_domain, as it is exported
    in ./xen/include/asm-arm/flushtlb.h. However, we restrict this
    function to be used only if altp2m is not active.