arm/p2m: Code movement in instr/data abort handlers
This commit moves code in the functions
"do_trap_data_(instr|abort)_guest" without changing the original
functionality. The code movement is limited to moving the struct npfec
out of the switch statements in both functions and the reordering of the
case statements FSC_FLT_PERM and FSC_FLT_TRANS in the function
"do_trap_data_abort_guest". Basically, this commit acts as a basis for
the following commit implementing the altp2m paging mechanism.

Signed-off-by: Sergej Proskurin <proskurin@sec.in.tum.de>
---
Cc: Stefano Stabellini <sstabellini@kernel.org>
Cc: Julien Grall <julien.grall@arm.com>