altp2m: Rename p2m_altp2m_check to altp2m_check
In this commit, we rename the function "p2m_altp2m_check" to "altp2m_check".
This is a preparation measure for the following commit which moves the renamed
functoin "altp2m_check" from p2m.c to altp2m.c in order to group all
altp2m-related functions to one spot (which is altp2m.c). The reason for
modifying the function's name is due the association of the function with the
associated .c file.

Signed-off-by: Sergej Proskurin <proskurin@sec.in.tum.de>
---
Cc:
---
v4: In this commit, we have pulled the renaming of the function
    "p2m_altp2m_check" out of the previous commit "altp2m: Introduce
    altp2m_switch_vcpu_altp2m_by_id"