arm/p2m: Introduce p2m_gva_to_ipa
The new function p2m_gva_to_ipa translates guest virtual into guest
physical addresses for a given p2m view. This is especially relevant in
the context of altp2m, as the ARM architecture only supports translation
of gva's into ipa's based on the active VTTBR. Therefore, the introduced
function switches to the VTTBR of the given p2m before translating the
gva; as soon as the translation is finished, p2m_gva_to_ipa restores the
old vttbr. That is, at runtime, while the active altp2m view might not
have the required gva mapping yet, this function switches the VTTBR of
the host's p2m (holding all valid mappings) to allow the translation to
succeed.

Signed-off-by: Sergej Proskurin <proskurin@sec.in.tum.de>
---
Cc:
---
v5: We have introduced this patch to our patch series.