x86/altp2m: Fix bug in vmx_vcpu_update_eptp - set EPTP_INDEX
The current implementation of the Xen altp2m subsystem suffers from a
bug that allows unprivileged guest domains that have access to the Xen
altp2m interface (with an active mode switch XEN_ALTP2M_mixed) to crash
not only the unprivileged guest, but also the Xen hypervisor itself.

This behavior can be deterministically reproduced by following the
following steps:
  (i)   HVMOP_altp2m_set_domain_state: enable altp2m
  (ii)  HVMOP_altp2m_vcpu_enable_notify: Register a #VE handler
  (iii) HVMOP_altp2m_set_domain_state: disable altp2m
  (iv)  HVMOP_altp2m_set_domain_state: enable altp2m
  (v)   HVMOP_altp2m_vcpu_enable_notify: Register a #VE handler

The last step (v) leads to a crash due to the fact that the current
implementation of the function vmx_vcpu_update_eptp does not update the
EPTP_INDEX corretly. This is because it only checks whether the flag
SECONDARY_EXEC_ENABLE_VIRT_EXCEPTIONS has been enabled for the
particular vCPU. Since this flag gets deactivated for every vCPU in
altp2m_vcpu_update_vmfunc_ve, the previously set value of INVALID_ALTP2M
remains in the EPTP_INDEX field of the VMCS. As such, the attempt of
re-initializing the #VE handler uses this wrong index (despite previous
initialization of the index 0 in vcpu_altp2m(v).p2midx) and crashes the
entire system.

This behavior can be either enforced by implementing a dedicated
hypercall or by (abusively) making use of the libxc library or via the
/dev/xen/privcmd device from inside the unprivileged guest domain.

This patch fixes the bug by generally checking the #VE support of the
CPU instead of the dynamically set capability by the vCPU.

Signed-off-by: Sergej Proskurin <proskurin@sec.in.tum.de>
---
Cc: Jun Nakajima <jun.nakajima@intel.com>
Cc: Kevin Tian <kevin.tian@intel.com>
Cc: Jan Beulich <jbeulich@suse.com>
Cc: Andrew Cooper <andrew.cooper3@citrix.com>
Cc: xen-devel@lists.xenproject.org
---